Improve the documentation
This sketches out how the documentation should go all the way from the
Apollo Project page down to the individual benchmark page.  This still
needs work as there is a bug whereby the main help page is the Theaceae
page rather than the Apollo page.  Copying over Camellia documentation
needs to be done as well.